+++ sentry-android-core/src/test/java/io/sentry/android/core/SentryAndroidTest.kt
@@ -490,6 +490,7 @@ class SentryAndroidTest {
           it is ActivityLifecycleIntegration ||
           it is ActivityBreadcrumbsIntegration ||
           it is UserInteractionIntegration ||
+          it is FeedbackShakeIntegration ||
           it is FragmentLifecycleIntegration ||
           it is SentryTimberIntegration ||
           it is AppComponentsBreadcrumbsIntegration ||